spectrum bus : rename beta.cpp to beta128.cpp as the original beta is somewhat different (nw)
