+++ src/devices/bus/spectrum/beta128.cpp
@@ -7,7 +7,7 @@
 *********************************************************************/
 
 #include "emu.h"
-#include "beta.h"
+#include "beta128.h"
 
 
 /***************************************************************************
